Header Beta v1.0
Fixed some of the issues with the header image, and text. Still an issue with the way I make the image a circle. (I.e. sometimes image will go out of div tag.) The temp fix will be a div tag that is half the size of the image. I figure that will be good enough for now.